Only allow a single ServiceWorker at / per domain.
Handling the lifetime of multiple ServiceWorkers with different scopes is
tricky at best. For example, just removing a ServiceWorker does not actually
remove it from clients. Explicit JS code is needed for this, but the delivery
of this code might in turn be intercepted by the old ServiceWorker.

To avoid these headaches, we restrict each domain to a single ServiceWorker at
the root scope ("/"). This way, each new version of the ServiceWorker (at the
fixed location /__sw.js) fully & atomically replaces previous registrations.
Any paths not explicitly handled by the ServiceWorker fall back to regular
fetch.